Refactor MarkdownToolbar.tsx
