+++ .gitignore
@@ -891,9 +891,24 @@ packages/app-mobile/components/NoteEditor/CodeMirror/webviewLogger.js.map
 packages/app-mobile/components/NoteEditor/EditLinkDialog.d.ts
 packages/app-mobile/components/NoteEditor/EditLinkDialog.js
 packages/app-mobile/components/NoteEditor/EditLinkDialog.js.map
-packages/app-mobile/components/NoteEditor/MarkdownToolbar.d.ts
-packages/app-mobile/components/NoteEditor/MarkdownToolbar.js
-packages/app-mobile/components/NoteEditor/MarkdownToolbar.js.map
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/MarkdownToolbar.d.ts
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/MarkdownToolbar.js
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/MarkdownToolbar.js.map
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/ToggleOverflowButton.d.ts
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/ToggleOverflowButton.js
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/ToggleOverflowButton.js.map
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/Toolbar.d.ts
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/Toolbar.js
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/Toolbar.js.map
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/ToolbarButton.d.ts
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/ToolbarButton.js
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/ToolbarButton.js.map
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/ToolbarOverflowRows.d.ts
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/ToolbarOverflowRows.js
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/ToolbarOverflowRows.js.map
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/types.d.ts
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/types.js
+packages/app-mobile/components/NoteEditor/MarkdownToolbar/types.js.map
 packages/app-mobile/components/NoteEditor/NoteEditor.d.ts
 packages/app-mobile/components/NoteEditor/NoteEditor.js
 packages/app-mobile/components/NoteEditor/NoteEditor.js.map
